New searches, new linux release, new Avalonia build

+++ Source/LoadByOS/LinuxConfigApp/LinuxInterop.cs
@@ -1,4 +1,5 @@
-﻿using LibationFileManager;
+﻿using AppScaffolding;
+using LibationFileManager;
 using System.Diagnostics;
 
 namespace LinuxConfigApp
@@ -24,12 +25,17 @@ namespace LinuxConfigApp
 		public void SetFolderIcon(string image, string directory) => throw new PlatformNotSupportedException();
 		public void DeleteFolderIcon(string directory) => throw new PlatformNotSupportedException();
 
+		public string ReleaseIdString => LibationScaffolding.ReleaseIdentifier.ToString() + (File.Exists("/bin/yum") ? "_RPM" : "");
+
 		//only run the auto upgrader if the current app was installed from the
-		//.deb package. Try to detect this by checking if the symlink exists.
+		//.deb or .rpm package. Try to detect this by checking if the symlink exists.
 		public bool CanUpgrade => Directory.Exists("/usr/lib/libation");
 		public void InstallUpgrade(string upgradeBundle)
 		{
-			RunAsRoot("apt", $"install '{upgradeBundle}'");
+			if (File.Exists("/bin/yum"))
+				RunAsRoot("yum", $"install '{upgradeBundle}'");
+			else
+				RunAsRoot("apt", $"install '{upgradeBundle}'");
 		}
 
 		public Process RunAsRoot(string exe, string args)
